#597 Moved sin and cos to angle object
