+++ eo-runtime/src/main/java/EOorg/EOeolang/EOmath/EOangle$EOcos.java
@@ -1,4 +1,4 @@
-package EOorg.EOeolang;
+package EOorg.EOeolang.EOmath;
 
 import org.eolang.AtComposite;
 import org.eolang.Data;
@@ -10,10 +10,10 @@ import org.eolang.XmirObject;
 /**
  * Cos.
  */
-@XmirObject(oname = "float.cos")
-public class EOfloat$EOcos extends PhDefault {
+@XmirObject(oname = "angle.cos")
+public class EOangle$EOcos extends PhDefault {
 
-    public EOfloat$EOcos(final Phi sigma) {
+    public EOangle$EOcos(final Phi sigma) {
         super(sigma);
         this.add("φ", new AtComposite(this, rho -> new Data.ToPhi(
             Math.cos(new Param(rho).strong(Double.class))
