+++ eo-runtime/src/main/java/EOorg/EOeolang/EOmath/EOangle$EOsin.java
@@ -1,4 +1,4 @@
-package EOorg.EOeolang;
+package EOorg.EOeolang.EOmath;
 
 import org.eolang.AtComposite;
 import org.eolang.Data;
@@ -10,10 +10,10 @@ import org.eolang.XmirObject;
 /**
  * Sin.
  */
-@XmirObject(oname = "float.sin")
-public class EOfloat$EOsin extends PhDefault {
+@XmirObject(oname = "angle.sin")
+public class EOangle$EOsin extends PhDefault {
 
-    public EOfloat$EOsin(final Phi sigma) {
+    public EOangle$EOsin(final Phi sigma) {
         super(sigma);
         this.add("φ", new AtComposite(this, rho -> new Data.ToPhi(
             Math.sin(new Param(rho).strong(Double.class))
