+++ eo-runtime/src/test/java/EOorg/EOeolang/EOmath/EOangleEOsinTest.java
@@ -1,5 +1,6 @@
-package EOorg.EOeolang;
+package EOorg.EOeolang.EOmath;
 
+import EOorg.EOeolang.EOmath.EOangle$EOsin;
 import org.eolang.Data;
 import org.eolang.Dataized;
 import org.hamcrest.MatcherAssert;
@@ -7,59 +8,59 @@ import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Test;
 
 /**
- * Test case for {@link EOfloat}
+ * Test case for {@link EOangle}
  */
-public final class EOfloatEOcosTest {
+public final class EOangleEOsinTest {
 
     @Test
-    public void cos0() {
-        final double cos0 = Math.cos(0);
+    public void sin0() {
+        final double sin0 = Math.sin(0);
         MatcherAssert.assertThat(
             new Dataized(
-                new EOfloat$EOcos(
+                new EOangle$EOsin(
                     new Data.ToPhi((double) 0)
                 )
             ).take(Double.class),
-            Matchers.equalTo(cos0)
+            Matchers.equalTo(sin0)
         );
     }
 
     @Test
-    public void cosPi() {
-        final double cosPi = Math.cos(Math.PI);
+    public void sinPi() {
+        final double sinPi = Math.sin(Math.PI);
         MatcherAssert.assertThat(
             new Dataized(
-                new EOfloat$EOcos(
+                new EOangle$EOsin(
                     new Data.ToPhi(Math.PI)
                 )
             ).take(Double.class),
-            Matchers.equalTo(cosPi)
+            Matchers.equalTo(sinPi)
         );
     }
 
     @Test
-    public void cosPiDiv2() {
-        final double cosPiDiv2 = Math.cos(Math.PI / 2);
+    public void sinPiDiv2() {
+        final double sinPiDiv2 = Math.sin(Math.PI / 2);
         MatcherAssert.assertThat(
             new Dataized(
-                new EOfloat$EOcos(
+                new EOangle$EOsin(
                     new Data.ToPhi(Math.PI / 2)
                 )
             ).take(Double.class),
-            Matchers.equalTo(cosPiDiv2)
+            Matchers.equalTo(sinPiDiv2)
         );
     }
 
     @Test
-    public void cosMinusPiDiv2() {
-        final double cosMinusPiDiv2 = Math.cos(-Math.PI / 2);
+    public void sinMinusPiDiv2() {
+        final double sinMinusPiDiv2 = Math.sin(-Math.PI / 2);
         MatcherAssert.assertThat(
             new Dataized(
-                new EOfloat$EOcos(
+                new EOangle$EOsin(
                     new Data.ToPhi(-Math.PI / 2)
                 )
             ).take(Double.class),
-            Matchers.equalTo(cosMinusPiDiv2)
+            Matchers.equalTo(sinMinusPiDiv2)
         );
     }
 }
